requirements.txt
# FIFRA Automation Requirements

# Browser Automation
selenium>=4.15.0

# Data Processing
pandas>=2.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Configuration
PyYAML>=6.0

# OCR (optional, for label verification fallback)
# pytesseract>=0.3.10
# pillow>=10.0.0

# Windows Automation (optional, for download dialogs)
# pyautogui>=0.9.54
